requirements.txt
gradio==5.33.0
gradio_client==1.10.2
groovy==0.1.2
safehttpx==0.1.6
fal-client==0.7.0
openai==1.54.4
langchain==0.2.16
langchain-openai==0.1.25
moviepy==1.0.3
opencv-python==4.10.0.84
numpy==1.26.4
Pillow==10.4.0
python-dotenv==1.0.1

# Character Consistency Dependencies
face-recognition==1.3.0
deepface==0.0.81
mediapipe==0.10.8
scikit-learn==1.3.2

# Face Enhancement Dependencies
gfpgan==1.3.8
realesrgan==0.3.0
basicsr==1.4.2
facexlib==0.3.0
torch>=2.0.0
torchvision>=0.15.0

# Additional utilities
dlib==19.24.2
opencv-contrib-python==4.10.0.84

# Face Swapping Dependencies
insightface==0.7.3
onnx==1.15.0
onnxruntime==1.16.3  # or onnxruntime-gpu for GPU support
